# api/openapi.yaml
openapi: 3.1.0

info:
  title: 3Default API
  version: 0.1.0

paths:
  /api/health:
    get:
      operationId: getHealth
      summary: Check application health
      responses:
        "200":
          description: Application is running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/ready:
    get:
      operationId: getReady
      summary: Check application readiness
      responses:
        "200":
          description: Application is ready to serve requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyResponse"
        "503":
          description: Application is not ready to serve requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyResponse"

  /api/projects:
    post:
      operationId: createProject
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          description: Invalid project input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthResponse:
      type: object
      additionalProperties: false
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ok

    CreateProjectRequest:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
        description:
          type:
            - string
            - "null"

    ProjectResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - visibility
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type:
            - string
            - "null"
        visibility:
          type: string
          enum:
            - private
            - public
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string

    ReadyResponse:
      type: object
      additionalProperties: false
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ready
            - unavailable